requirements.txt
Flask==1.0
numpy==1.17.2
pandas==0.25.1
psycopg2-binary==2.8.3
